refactor(scripts): extract _runner_common.py — dedupe flux2/z-image shared helpers
`scripts/flux2_macos.py` and `scripts/z_image_turbo.py` carried ~200 LOC
of byte-for-byte-duplicated helpers: `pick_device`, `make_generator`,
`apply_memory_optimizations`, `write_sidecar`, `make_stepwise_callback`,
`_emit_user_error`, `_repo_from_hf_error`, and the entire bottom-of-file
HF cause-chain walker.

Pulled into a new `scripts/_runner_common.py` (mirrors the existing
`lora_utils.py` extraction). `make_stepwise_callback` takes an optional
`unpack_latents` callable so the flux2 packed-transformer-latent path
stays runner-specific while the rest of the decode is shared.
`install_hf_error_handler` is a `@decorator` over `main()` that owns
the gated/notfound/401 dispatch — the trailing `if __name__ ==
"__main__":` block in each runner collapses to a single `main()` call.

Smoke: both runners' `--help` parses cleanly; server test suite
4012/4012 green.
