+++ scripts/flux2_macos.py
@@ -25,8 +25,6 @@ import inspect
 import json
 import os
 import sys
-import threading
-from contextlib import contextmanager
 from pathlib import Path
 
 # Must precede `import torch` — enables the fast-math kernel path on MPS,
@@ -37,54 +35,16 @@ import torch
 from PIL import Image
 
 sys.path.insert(0, str(Path(__file__).resolve().parent))
+from _runner_common import (  # noqa: E402
+    apply_memory_optimizations,
+    heartbeat,
+    install_hf_error_handler,
+    make_generator,
+    make_stepwise_callback,
+    pick_device,
+    write_sidecar,
+)
 from lora_utils import apply_loras  # noqa: E402
-
-
-@contextmanager
-def heartbeat(stage: str, interval: float = 20.0):
-    # Diffusers' from_pretrained on a fully-cached 10-25 GB model is silent
-    # for several minutes (mmap + weight assignment, no tqdm), which trips the
-    # JS-side idle watchdog (default 5min). Emit a non-noise stderr line every
-    # `interval` seconds so handleLine() in imageGen/local.js sees activity
-    # and resets lastActivityAt. True hangs (GIL-pinned C extension, no I/O)
-    # still trip the watchdog because the heartbeat thread can't print either.
-    stop = threading.Event()
-
-    def beat():
-        elapsed = 0
-        while not stop.wait(interval):
-            elapsed += int(interval)
-            print(f"STAGE:{stage}:heartbeat:{elapsed}s", file=sys.stderr, flush=True)
-
-    t = threading.Thread(target=beat, daemon=True)
-    t.start()
-    try:
-        yield
-    finally:
-        stop.set()
-        t.join(timeout=interval + 1)
-
-
-def pick_device(requested: str) -> str:
-    if requested == "auto":
-        if torch.backends.mps.is_available():
-            return "mps"
-        if torch.cuda.is_available():
-            return "cuda"
-        return "cpu"
-    if requested == "mps" and not torch.backends.mps.is_available():
-        print("⚠️ MPS requested but unavailable — falling back to CPU", file=sys.stderr)
-        return "cpu"
-    if requested == "cuda" and not torch.cuda.is_available():
-        print("⚠️ CUDA requested but unavailable — falling back to CPU", file=sys.stderr)
-        return "cpu"
-    return requested
-
-
-def make_generator(device: str, seed: int) -> torch.Generator:
-    if device in ("cuda", "mps"):
-        return torch.Generator(device).manual_seed(int(seed))
-    return torch.Generator().manual_seed(int(seed))
 
 
 def probe_hf_auth(repo: str) -> None:
@@ -207,18 +167,6 @@ def load_pipeline_int8(repo: str, base_repo: str, device: str, dtype):
     return pipe
 
 
-def apply_memory_optimizations(pipe) -> None:
-    if hasattr(pipe, "enable_attention_slicing"):
-        pipe.enable_attention_slicing()
-    if hasattr(pipe, "enable_vae_slicing"):
-        pipe.enable_vae_slicing()
-    vae = getattr(pipe, "vae", None)
-    if hasattr(pipe, "enable_vae_tiling"):
-        pipe.enable_vae_tiling()
-    elif vae is not None and hasattr(vae, "enable_tiling"):
-        vae.enable_tiling()
-
-
 def _unpack_flux2_latents(latents, height: int, width: int, vae_scale: int = 8, patch: int = 2):
     """Convert Flux2 transformer-packed latents `(B, num_patches, C*P*P)` back
     to the VAE-friendly `(B, C, H_lat, W_lat)` layout.
@@ -240,74 +188,6 @@ def _unpack_flux2_latents(latents, height: int, width: int, vae_scale: int = 8, 
     latents = latents.view(bsz, h, w, c, patch, patch)
     latents = latents.permute(0, 3, 1, 4, 2, 5)
     return latents.reshape(bsz, c, h * patch, w * patch)
-
-
-def make_stepwise_callback(stepwise_dir: str, pipe, height: int, width: int):
-    """Return a `callback_on_step_end` that decodes the running latent into a
-    small preview PNG. local.js's `processLatestFrame` watches this dir and
-    streams the freshest frame to the SSE client."""
-
-    if not stepwise_dir:
-        return None
-    out = Path(stepwise_dir)
-    out.mkdir(parents=True, exist_ok=True)
-    vae = pipe.vae
-    scaling = getattr(vae.config, "scaling_factor", None)
-    if scaling is None:
-        scaling = 1.0
-    shift = getattr(vae.config, "shift_factor", None)
-    if shift is None:
-        shift = 0.0
-    # Capture VAE weight dtype + device so the callback can align the scaled
-    # latents before decode — pipelines sometimes keep latents in float32
-    # while the VAE is bfloat16 (which errors with "Input type (float) and
-    # bias type (c10::BFloat16) should be the same").
-    try:
-        vae_param = next(vae.parameters())
-        vae_dtype, vae_device = vae_param.dtype, vae_param.device
-    except StopIteration:
-        vae_dtype, vae_device = None, None
-
-    fired = {"count": 0, "saved": 0}
-
-    @torch.no_grad()
-    def cb(pipe, step_index, _timestep, callback_kwargs):
-        fired["count"] += 1
-        latents = callback_kwargs.get("latents")
-        if latents is None:
-            if step_index == 0:
-                print("⚠️ stepwise: latents missing from callback_kwargs", file=sys.stderr)
-            return callback_kwargs
-        if step_index == 0:
-            print(f"🖼️  stepwise: callback live, latents.shape={tuple(latents.shape)}", file=sys.stderr)
-        # Best-effort decode. Errors here must not abort generation — the
-        # final image is still produced after the last step.
-        try:
-            # Flux2 latents arrive packed (B, num_patches, C*p*p). VAE expects
-            # (B, C, H, W). Skip unpack if it's already 4-D (future-proof).
-            if latents.dim() == 3:
-                latents = _unpack_flux2_latents(latents, height, width)
-            scaled = latents / scaling + shift
-            if vae_dtype is not None and (scaled.dtype != vae_dtype or scaled.device != vae_device):
-                scaled = scaled.to(device=vae_device, dtype=vae_dtype)
-            decoded = vae.decode(scaled, return_dict=False)[0]
-            decoded = (decoded.clamp(-1, 1) + 1) / 2
-            arr = (decoded[0].float().cpu().permute(1, 2, 0).numpy() * 255).astype("uint8")
-            img = Image.fromarray(arr)
-            img.thumbnail((512, 512), Image.LANCZOS)
-            img.save(out / f"step_{step_index + 1}.png", "PNG", optimize=False)
-            fired["saved"] += 1
-        except Exception as err:
-            print(f"⚠️ stepwise preview failed at step {step_index}: {type(err).__name__}: {err}", file=sys.stderr)
-        return callback_kwargs
-
-    cb._stats = fired
-    return cb
-
-
-def write_sidecar(output: str, payload: dict) -> None:
-    sidecar = Path(output).with_suffix(".metadata.json")
-    sidecar.write_text(json.dumps(payload, indent=2))
 
 
 def parse_args() -> argparse.Namespace:
@@ -335,6 +215,7 @@ def parse_args() -> argparse.Namespace:
     return p.parse_args()
 
 
+@install_hf_error_handler
 def main() -> None:
     args = parse_args()
 
@@ -369,7 +250,13 @@ def main() -> None:
             (int(args.width), int(args.height)), Image.LANCZOS
         )
 
-    callback = make_stepwise_callback(args.stepwise_image_output_dir, pipe, int(args.height), int(args.width))
+    callback = make_stepwise_callback(
+        args.stepwise_image_output_dir,
+        pipe,
+        int(args.height),
+        int(args.width),
+        unpack_latents=_unpack_flux2_latents,
+    )
     # Flux2KleinPipeline.__call__ doesn't always accept negative_prompt or
     # strength — passing an unsupported kwarg raises TypeError. Filter to
     # what the live signature actually accepts.
@@ -449,84 +336,5 @@ def main() -> None:
     print(f"✅ flux2 saved {args.output} (seed={seed})", file=sys.stderr)
 
 
-def _emit_user_error(kind: str, message: str, repo: str = "") -> None:
-    """Emit a structured single-line error the server's stderr parser picks up
-    for the SSE error event. `kind` is a stable identifier the UI maps to a
-    friendly heading; `message` is the human prose; `repo` (optional) is the
-    HF repo to deep-link the user to so they can request access / check token."""
-    line = f"USER_ERROR:{kind}"
-    if repo:
-        line += f":{repo}"
-    print(line, file=sys.stderr, flush=True)
-    print(f"❌ {message}", file=sys.stderr, flush=True)
-
-
 if __name__ == "__main__":
-    try:
-        main()
-    except KeyboardInterrupt:
-        sys.exit(130)
-    except SystemExit:
-        raise
-    except Exception as err:
-        # Top-level safety net — every error type we recognise gets a friendly
-        # USER_ERROR: line so the UI can surface "request access at <link>"
-        # instead of bare "Exit code 1". Anything we don't recognise still
-        # falls through to a printed traceback above + a generic USER_ERROR
-        # below, so the user sees *something* actionable.
-        from huggingface_hub.errors import GatedRepoError, RepositoryNotFoundError, HfHubHTTPError
-        # Walk the cause chain — diffusers wraps HF errors in OSError, so the
-        # innermost exception is what tells us the real story.
-        chain = []
-        cur = err
-        while cur is not None:
-            chain.append(cur)
-            cur = cur.__cause__ or cur.__context__
-        def _repo_from_hf_error(hf_err):
-            """huggingface_hub doesn't always populate `repo_id` on the error.
-            Fall back to parsing the failing URL — `/<owner>/<repo>/resolve/...`
-            for file downloads or `/api/models/<owner>/<repo>` for API hits."""
-            repo = getattr(hf_err, "repo_id", None) or ""
-            if repo:
-                return repo
-            url = getattr(getattr(hf_err, "response", None), "url", None) or \
-                  getattr(getattr(hf_err, "request", None), "url", None)
-            if url is None:
-                return ""
-            path = str(url).split("huggingface.co", 1)[-1].lstrip("/")
-            parts = path.split("/")
-            if parts[:1] == ["api"] and len(parts) >= 4 and parts[1] in {"models", "datasets", "spaces"}:
-                return f"{parts[2]}/{parts[3]}"
-            if len(parts) >= 2 and parts[0] not in {"api", "settings", "join"}:
-                return f"{parts[0]}/{parts[1]}"
-            return ""
-
-        gated = next((e for e in chain if isinstance(e, GatedRepoError)), None)
-        notfound = next((e for e in chain if isinstance(e, RepositoryNotFoundError)), None)
-        http = next((e for e in chain if isinstance(e, HfHubHTTPError)), None)
-        if gated is not None:
-            repo = _repo_from_hf_error(gated)
-            url = f"https://huggingface.co/{repo}" if repo else "https://huggingface.co/"
-            _emit_user_error(
-                "gated_repo",
-                f"Access to {repo or 'the model repo'} is restricted. Visit {url} "
-                f"to request access, then make sure your HF token is set in PortOS.",
-                repo,
-            )
-            sys.exit(2)
-        status = getattr(getattr(http, "response", None), "status_code", None) if http else None
-        if status == 401:
-            _emit_user_error(
-                "hf_unauthorized",
-                "HuggingFace rejected the token (401). Check that the token is valid "
-                "and has read access, then re-paste it in PortOS.",
-            )
-            sys.exit(2)
-        if notfound is not None:
-            repo = _repo_from_hf_error(notfound)
-            _emit_user_error("repo_not_found", f"HF repo not found: {repo or '(unknown)'}", repo)
-            sys.exit(2)
-        # Unknown failure — emit the original traceback (already printed above)
-        # plus a generic structured marker so the UI shows something useful.
-        _emit_user_error("unknown", f"{type(err).__name__}: {err}")
-        raise
+    main()
